airline_multiagent_project_yandex_patch/airline_multiagent/requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
pydantic>=2.7.0

python-dotenv>=1.0.1
